+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## **[Unreleased]**
 ### Changed
 - Updated dependency `whoami` to 1.1.5
+- Support reusing login tokens when switching registries: old login tokens are now not cleared automatically
+- Add `wapm login --user <NAME> --password <PASSWORD>`
+- Add `wapm init <projectname>` to make initializing projects quicker
 
 ## [0.5.1] - 2021-03-30
 ### Added